tests/wpt/web-platform-tests/custom-elements/instantiating-custom-elements/custom-element-type-local-name.html

<!DOCTYPE html>
<html>
<head>
<title>Instantiation of custom element: custom element type is given via the local name of the custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="assert" content="The custom element type is given to a custom element at the time of its instantation in one of the two ways: 1. As the local name of the custom element.">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    var doc = newHTMLDocument();
    var GeneratedConstructor = doc.registerElement('x-a');
    doc.body.innerHTML = '<x-a id="x-a"></x-a>';
    var customElement = doc.querySelector('#x-a');
    assert_equals(Object.getPrototypeOf(customElement), GeneratedConstructor.prototype,
        'Custom element type should be the type, specified by the local name of ' +
        'the custom element');
}, 'Test custom element type, which is given via the local name of the custom element. ' +
    'Custom element created via innerHTML property');


testInIFrame('../resources/x-element.html', function(doc) {
    var GeneratedConstructor = doc.registerElement('x-element');
    var xelement = doc.querySelector('#x-element');
    assert_equals(Object.getPrototypeOf(xelement), GeneratedConstructor.prototype,
        'Custom element type should be the type, specified by the local name of ' +
        'the custom element');
}, 'Test custom element type, which is given via the local name of the custom element. ' +
    'Custom element is defined in loaded HTML document');
</script>
</body>
</html>
